refactor: add Button class with render functionality
